% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatch.histogram.R
\name{adductMatch.histogram}
\alias{adductMatch.histogram}
\title{Find closest-matching mass difference bin for known adducts}
\usage{
\method{adductMatch}{histogram}(x, add = adducts, density = FALSE)
}
\arguments{
\item{x}{histogram object produced by \code{\link{hist.massdiff}}}

\item{add}{data.frame of adduct masses (default: "adducts" dataset in package)}

\item{density}{logical; return density instead of counts? (default: FALSE)}
}
\value{
data.frame of adducts with counts/density from closest-matching bin
        and their corresponding quantiles in mass difference bins
}
\description{
For use with internal dataset "adducts" and other similar data
}
\seealso{
\code{\link{topAdducts}} to find the closest-matching adducts for
         the mass differences with the highest counts (the converse of the
         current function)
}
