% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatch.massdiffhist.R
\name{adductMatch.massdiffhist}
\alias{adductMatch.massdiffhist}
\title{Report closest-matching adduct type for a mass difference histogram}
\usage{
\method{adductMatch}{massdiffhist}(x, add = adducts, density = FALSE)
}
\arguments{
\item{x}{massdiffhist object produced by \code{\link{hist.massdiff}}}

\item{add}{data.frame of adduct masses (default: "adducts" dataset in package)}

\item{density}{logical; return density instead of counts? (default: FALSE)}
}
\value{
data.frame of adducts with counts/density from closest-matching bin
        and their corresponding quantiles in mass difference bins
}
\description{
Given a histogram of mass differences (from applying \code{hist} on a massdiff
object), report the bins most closely matching a list of potential molecular
transformations (adduct types). The built-in data sets \code{\link{adducts}}
and \code{\link{adducts2}} are examples of such lists of potential adducts.
This function is designed for a quick exploratory overview of an MSI data set,
especially when the number of peaks is large. To report each peak pair that
has a match to the list of potential adducts, apply \code{\link{adductMatch}}
directly to a massdiff object.
}
\seealso{
\code{\link{adductMatch.massdiff}} reports closest-matching adducts
         for each individual peak pair.

\code{\link{topAdducts}} to find the closest-matching adducts for
         the mass differences with the highest counts (the converse of the
         current function)
}
