% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatchDiffs.R
\name{adductMatchDiffs}
\alias{adductMatchDiffs}
\title{Find closest-matching known adducts for a set of mass differences}
\usage{
adductMatchDiffs(diff, add = adducts, width = 0.001)
}
\arguments{
\item{diff}{data.frame of mass differences produced by \code{\link{diffTabulate}}}

\item{add}{data.frame of adduct masses (default: "adducts" dataset in package)}

\item{width}{numeric; the precision to which to match the mass value (default: 0.01)}
}
\value{
data.frame of mass differences given to argument diff, with an
        additional column reporting the closest matches
}
\description{
For use with internal dataset "adducts" and other similar data
}
\seealso{
\code{\link{topAdducts}} to find the closest-matching adducts for
         the mass differences with the highest counts (the converse of the
         current function)
}
