% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adducts}
\alias{adducts}
\title{Common mass spectrometry adducts}
\format{A data.frame with 92 rows and 3 columns:
\describe{
  \item{name}{Common name of adduct}
  \item{formula}{Formula of adduct}
  \item{mass}{Mass of adduct in m/z units}
}}
\source{
Moritz Janda, Manuel Liebeke
}
\usage{
adducts
}
\description{
A dataset containing the masses and formulae of selected adducts commonly
observed in mass spectrometry imaging data.
}
\keyword{datasets}
