% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adducts2}
\alias{adducts2}
\title{Shorter list of mass spectrometry adducts}
\format{A data.frame with 9 rows and 3 columns:
\describe{
  \item{name}{Common name of adduct}
  \item{formula}{Formula of adduct}
  \item{mass}{Mass of adduct in m/z units}
}}
\source{
Moritz Janda, Manuel Liebeke
}
\usage{
adducts2
}
\description{
A dataset containing the masses and formulae of selected adducts commonly
observed in mass spectrometry imaging data, subset of the list \code{adducts}
}
\keyword{datasets}
