% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeIntensityCutoffsCumul.R
\name{analyzeIntensityCutoffsCumul}
\alias{analyzeIntensityCutoffsCumul}
\title{Calculate effects of applying cumulative intensity-level cutoffs on mass spectrum}
\usage{
analyzeIntensityCutoffsCumul(df, by = "intensities", value = NULL, pc = 1,
  plot = TRUE, log.plot = TRUE, report.peaks = FALSE)
}
\arguments{
\item{df}{data.frame of three columns: peaks, counts, intensities, produced
by function \code{\link{sumPeakIntensities}}, or by external script
msicsvprofiler.pl if matrix is too large to import into R}

\item{by}{string; Apply cutoff by intensities or counts (default: intensities)}

\item{value}{numeric; Minimum intensity/counts cutoff to apply}

\item{pc}{numeric; Percent intensity/counts cutoff to apply (ignored if value
supplied). For example, 50% would find the top peaks arranged by
intensity, whose combined total intensity accounts for 50% of the total
intensity in the data (Default: 1)}

\item{plot}{logical; Plot distribution? (default: TRUE)}

\item{log.plot}{logical; Plot cumulative values with log scale (default: TRUE)}

\item{report.peaks}{logical: Report list of peaks above the cutoff, as vector
(default: FALSE)}
}
\description{
Given a table of total intensities per mass peak, produced by function
\code{\link{sumPeakIntensities}}, calculate the number of peaks retained
if a minimum intensity cutoff is applied (either as absolute value or as
percentage of the TOTAL intensity value observed) to the cumulative sum of
the intensity values. This has similarities to the N50 metric applied to genome
assemblies.
}
\seealso{
\code{\link{analyzeIntensityCutoffsDistr}} does a similar analysis
         but without taking cumulative sums, instead using the empirical CDF
         of the values.
}
