% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal2msimat.MSContinuousImagingExperiment.R
\name{cardinal2msimat.MSContinuousImagingExperiment}
\alias{cardinal2msimat.MSContinuousImagingExperiment}
\title{Convert Cardinal MSContinuousImagingExperiment object to msimat object}
\usage{
\method{cardinal2msimat}{MSContinuousImagingExperiment}(d)
}
\arguments{
\item{d}{MSContinuousImagingExperiment; MSI data processed by Cardinal, must
be processed with peakBin.}
}
\description{
Convert imaging data in MSContinuousImagingExperiment, which is stored
internally as a sparse matrix (class sparse_matc object) to an msimat object.
}
