% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal2msimat.MSProcessedImagingExperiment.R
\name{cardinal2msimat.MSProcessedImagingExperiment}
\alias{cardinal2msimat.MSProcessedImagingExperiment}
\title{Convert Cardinal MSProcessedImagingExperiment object to msimat object}
\usage{
\method{cardinal2msimat}{MSProcessedImagingExperiment}(d)
}
\arguments{
\item{d}{MSProcessedImagingExperiment; MSI data processed by Cardinal, must
be processed with peakBin.}
}
\description{
Convert imaging data in MSProcessedImagingExperiment, which is stored
internally as a sparse matrix (class sparse_matc object) to an msimat object.
}
