% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal2msimat.R
\name{cardinal2msimat}
\alias{cardinal2msimat}
\title{Convert Cardinal objects to msimat object}
\usage{
cardinal2msimat(d)
}
\arguments{
\item{d}{MSProcessedImagingExperiment; MSI data in Cardinal format, must
be already pre-processed with peakBin.}
}
\value{
Object of class msimat
}
\description{
Convert imaging data imported with Cardinal v2+ to msimat format for analysis
with mass2adduct. The data must be already pre-processed and peak-binned.
}
\seealso{
\code{\link{msimat}}
}
