% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comb2M}
\alias{comb2M}
\title{C++ version of combn function}
\usage{
comb2M(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric matrix with three rows: all pairwise combinations of values
        in vector x (rows 1 and 2), and their arithmetic difference (row 3)
}
\description{
C++ version of combn function
}
