% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPairsMSI.msitsm.R
\name{corrPairsMSI.msitsm}
\alias{corrPairsMSI.msitsm}
\title{Calculate correlations of pairs of mass peaks from MSI data}
\usage{
\method{corrPairsMSI}{msitsm}(d, pairs, p.val = 0.05, method = "pearson",
  alternative = "greater", ...)
}
\arguments{
\item{d}{data.frame; MSI data with peaks as columns and pixels as rows,
output from \code{\link{readMSI}} function}

\item{pairs}{data.frame; Data frame with pairs of peaks (as rows) for which
to calculate correlations.}

\item{p.val}{numeric; p-value cutoff (before Bonferroni correction) (default: 0.05)}

\item{...}{Other parameters to pass to \code{cor.test}}
}
\value{
Object of class data.frame with the following fields:

        label - Label for row

        parental_ion - First peak mass in each pair

        adduct_ion - Second peak mass in the pair

        Estimate - Estimated correlation

        P.value - P-value for the correlation

        Significance - Whether p-value meets cutoff (specified by "p.val", with
                       Bonferroni correction)
}
\description{
Calculate correlations of pairs of mass peaks from MSI data (imported to
R with the \code{\link{readMSI}} function). The list of pairs is supplied
as a data.frame with the parameter \code{pairs}.
}
\details{
Example usage scenario for this function: Mass differences for all pairwise
combinations of masses are tabulated with \code{\link{diffTabulate}}, and the
peak pairs corresponding to a specific adduct of interest are extracted with
\code{\link{diffGetPeaks}}. Check which peaks are significantly correlated to
each other with this function.

Correlation is calculated with \code{\link{cor.test}} function from the
\code{stats} package, and by default uses the Pearson methone (one-sided, i.e.
positive correlations). The Bonferroni correction is applied to the p-values
before assessing significance, but the original p-value is reported in column
P.value.
}
\seealso{
\code{\link{corrSinglePeakMSI}} to calculate correlations of all
         peaks in a dataset vs. a single peak of interest
}
