% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPairsMSIchunks.R
\name{corrPairsMSIchunks}
\alias{corrPairsMSIchunks}
\title{Calculate correlations of pairs of mass peaks from MSI data in chunks}
\usage{
corrPairsMSIchunks(d, diff, p.val = 0.05, method = c("pearson", "kendall",
  "spearman"), alternative = c("two.sided", "less", "greater"),
  ncores = NULL, mem.limit = 5, ...)
}
\arguments{
\item{d}{msimat; MSI data with peaks as columns and pixels as rows,
output from \code{\link{readMSI}} function}

\item{diff}{massdiff; List of mass differences, parent and putative adduct
ions, as produced by function \code{\link{diffTabulate}}}

\item{p.val}{numeric; p-value cutoff (before Bonferroni correction) (default:
0.05)}

\item{method}{string; Method to use for \code{\link{cor.test}} (default:
"pearson")}

\item{alternative}{string; Which alternative for \code{\link{cor.test}}
(default: "greater")}

\item{ncores}{integer; Number of cores if using parallel version. Default is
total number of cores minus one.}

\item{mem.limit}{integer; Memory limit estimate (in Gb). (default: 5)}

\item{...}{Other parameters to pass to \code{\link{cor.test}}}
}
\value{
Object of class data.frame with the following fields:

        A - First peak mass in each pair

        B - Second peak mass in the pair

        Estimate - Estimated correlation

        P.value - P-value for the correlation

        Significance - Whether p-value meets cutoff (specified by "p.val", with
                       Bonferroni correction)
}
\description{
Calculate correlations of pairs of mass peaks from MSI data (imported to
R with the \code{\link{readMSI}} function). The list of pairs is supplied
as a data.frame with the parameter \code{pairs}, but is broken up into
"chunks" for processing, to avoid going over a specified memory limit.
}
\details{
This function is otherwise equivalent to \code{\link{corrPairsMSI}} with
option \code{how="parallel"}. It uses forking processes and is therefore not
suitable for Windows systems. The memory requirement is estimated
conservatively but has only been tested empirically (i.e. the process might
go over the limit in practice).

Example usage scenario for this function: Mass differences for all pairwise
combinations of masses are tabulated with \code{\link{diffTabulate}}, and the
peak pairs corresponding to a specific adduct of interest are extracted with
\code{\link{diffGetPeaks}}. Check which peaks are significantly correlated to
each other with this function.

Correlation is calculated with \code{\link{cor.test}} function and by default
uses the Pearson method (two-sided, because correlations can be both positive
and negative). The Bonferroni correction is applied to the p-values before
assessing significance, but the original p-value is reported in column
"P.value" and can be used e.g. for false discovery rate analysis.
}
\seealso{
\code{\link{corrPairsMSI}} function which does not split input into
         chunks.
}
