% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSinglePeakMSI.R
\name{corrSinglePeakMSI}
\alias{corrSinglePeakMSI}
\title{Calculate correlations of single target ion to other mass peaks from MSI data}
\usage{
corrSinglePeakMSI(d, peak = NULL, p.val = 0.05, ...)
}
\arguments{
\item{d}{data.frame; MSI data with peaks as columns and pixels as rows,
output from readMSI() function}

\item{peak}{numeric; Target ion for which to calculate correlations vs other ions
in MSI data (input d).}

\item{p.val}{numeric; p-value cutoff (before Bonferroni correction) (default: 0.05)}

\item{...}{Other parameters to pass to \code{cor.test}}
}
\value{
Object of class data.frame with the following fields:

        label - Label for row

        ion_of_interest - Target ion peak mass specified with parameter "peak"

        all_ions - Peak masses for other ions in MSI data "d"

        Estimate - Estimated correlation

        P.value - P-value for the correlation (without Bonferroni correction)

        Significance - Whether p-value meets cutoff (specified by "p.val", with
                       Bonferroni correction)
}
\description{
Calculate correlations of a target ion mass to all other mass peaks from
MSI data (imported to R with the \code{\link{readMSI}} function). The target
mass is supplied with the parameter \code{peak} and must match one of the
mass values in the MSI dataset supplied with parameter \code{d}.
}
\details{
Correlation is calculated with \code{\link{cor.test}} function from the
\code{stats} package, and uses the default two-sided Pearson method. The
Bonferroni correction is applied to the p-values before assessing significance,
but the original p-value is reported in column P.value.
}
\seealso{
\code{\link{corrPairsMSI}} to calculate correlations of pairs of ions
         e.g. the output from \code{\link{diffTabulate}}
}
