% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPeaksFromDiff.R
\name{countPeaksFromDiff}
\alias{countPeaksFromDiff}
\title{Count peaks corresponding to putative adduct types}
\usage{
countPeaksFromDiff(diff, which = c("adduct", "parent"))
}
\arguments{
\item{diff}{Object of class massdiff that has been annotated with adduct
types by \code{\link{adductMatch}} function.}

\item{which}{Whether to enumerate putative adducts or parent ions}
}
\value{
data.frame with number of peaks corresponding to each putative
        chemical transformation type.
}
\description{
Adduct matching with \code{\link{adductMatch}} or visualizing the massdiff
histogram with \code{\link{hist.massdiff}} will give an overview of the
number of ion pairs corresponding to transformations of interest. However,
this does not directly indicate the number of actual ions that participate in
the transformations, because a given ion can participate in multiple
transformations either as parent or adduct.
}
\details{
This function takes an annotated massdiff result, where adducts of interest
have already been matched to the mass difference values, and tabulates the
numbers of parent or adduct ions per adduct type. The total may sum to more
than 100%, because a given ion may represent more than one type of
chemical transformation.
}
