% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffGetPeaks.R
\name{diffGetPeaks}
\alias{diffGetPeaks}
\title{Find pairs of mass peaks corresponding to a specific mass difference}
\usage{
diffGetPeaks(d, mass = NULL, width = 0.01)
}
\arguments{
\item{d}{data.frame; output from diffTabulate}

\item{mass}{numeric; mass difference in m/z units}

\item{width}{numeric; range of values to use}
}
\value{
data.frame of peak pairs where the difference is within the
width of the specified mass of interest
}
\description{
Find pairs of mass peaks corresponding to a specific mass difference
}
