% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffHist.R
\name{diffHist}
\alias{diffHist}
\title{Histogram of mass difference list}
\usage{
diffHist(d, widthFunc = "equal", width = 0.01, plot = FALSE, ...)
}
\arguments{
\item{d}{data.frame; three numeric columns (A, B, diff) representing mass
peaks and their differences}

\item{widthFunc}{character; function to use to bin mass differences into
histogram. (default "equal", other options to be added)}

\item{width}{numeric; bin width in m/z units (default 0.01)}

\item{plot}{logical; whether histogram graphic should be plotted (default FALSE)}

\item{...}{Parameters to be passed to hist()}
}
\value{
Object of class hist
}
\description{
Histogram of mass difference list
}
\seealso{
\code{\link{diffTabulate}} to generate the mass difference list
}
