% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTabulate.msidf.R
\name{diffTabulate.msidf}
\alias{diffTabulate.msidf}
\title{Tabulate pairwise combinations of mass peaks from imported MSI data}
\usage{
\method{diffTabulate}{msidf}(d)
}
\arguments{
\item{d}{data.frame; MS imaging data imported with \code{readMSI}}

\item{override.limit}{logical; continue with computation even if vector
length is >5000 (default: FALSE)}
}
\value{
data.frame of all pairs of masses and their respective differences
}
\description{
Takes all pairwise combinations of mass peaks from MS imaging data imported
with the function \code{\link{readMSI}}.
}
\details{
The number of possible pairs grows combinatorially and so the calculation
can take a very long time for more than 5000 mass values. Use the
\code{override.limit} parameter to override the default limit of 5000.
}
\seealso{
\code{\link{diffTabulate}} for tabulating from a simple numeric
         list of masses, \code{\link{diffHist}} to plot histogram of mass
         differences, \code{\link{readMSI}} to read MS imaging data
}
