% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTabulate.msitsm.R
\name{diffTabulate.msitsm}
\alias{diffTabulate.msitsm}
\title{Tabulate pairwise combinations of mass peaks from imported MSI data}
\usage{
\method{diffTabulate}{msitsm}(d)
}
\arguments{
\item{d}{numeric vector representing m/z values of mass peaks}
}
\value{
data.frame of all pairs of masses and their respective differences
}
\description{
Takes all pairwise combinations of mass peaks from MS imaging data imported
with the function \code{\link{readMSI}}.
}
\details{
The number of possible pairs grows combinatorially and the calculation will
fail for >65536 values (the function uses a 16-bit integer internally).
}
\seealso{
\code{\link{diffHist}} to plot histogram of mass differences
}
