% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTabulate.R
\name{diffTabulate}
\alias{diffTabulate}
\title{Tabulate pairwise combinations of mass peaks}
\usage{
diffTabulate(d)
}
\arguments{
\item{d}{Either an MSI data set of class msimat, imported with the function
\code{\link{readMSI}}, or a numeric vector representing m/z values
of mass peaks}
}
\value{
Object of class \code{massdiff} with three elements:

                   A - m/z values for peak pairs; for a given pair, A will
                       be lower than B. Requries that the peak list in the
                       input be sorted ascending numerically by mass.

                   B - m/z values for peak pairs

                   diff - their respective differences
}
\description{
Takes all pairwise combinations of masses from either an MSI data set that
was imported with the function \code{\link{readMSI}}, or from a simple
numeric vector of masses.
}
\details{
The number of possible pairs grows combinatorially and the calculation will
fail for >65536 values (the function uses a 16-bit integer internally).
}
\seealso{
\code{\link{diffHist}} to plot histogram of mass differences

\code{\link{adductMatch}} to match mass differences to possible
         molecular transformations

\code{\link{corrPairsMSI}} to test whether given peak pairs are
         correlated, using MSI intensity data from massdiff and msimat objects
}
