% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPeaks.data.frame.R
\name{filterPeaks.data.frame}
\alias{filterPeaks.data.frame}
\title{Filter peak list of MSI data}
\usage{
\method{filterPeaks}{data.frame}(d, how = c("topX", "XofTop", "XofTotal"),
  x = NULL, index = FALSE)
}
\description{
Filter peak list of MSI data
}
