% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPeaks.msimat.R
\name{filterPeaks.msimat}
\alias{filterPeaks.msimat}
\title{Filter peak list of MSI data}
\usage{
\method{filterPeaks}{msimat}(d, how = c("topX", "XofTop", "XofTotal"),
  x = NULL, index = NULL)
}
\description{
Filter peak list of MSI data
}
