% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histWeighted.R
\name{histWeighted}
\alias{histWeighted}
\title{Correlation-weighted histogram of massdiff object}
\usage{
histWeighted(diff, widthFunc = "equal", width = 0.01, ncores = NULL,
  ...)
}
\arguments{
\item{diff}{massdiff; Output from function \code{\link{massdiff}},
containing three numeric vectors (A, B, diff) representing mass
peaks and their differences, processed with \code{\link{corrPairsMSI}}
which adds columns Estimate, P.value, and Significance}

\item{widthFunc}{character; function to use to bin mass differences into
histogram. (default "equal", other options to be added)}

\item{width}{numeric; bin width in m/z units (default 0.01)}

\item{ncores}{integer; number of processors to use for parallelized
implementation. If NULL (default) then run with single processor}

\item{...}{Other options to be passed to hist()}
}
\value{
Object of classes hist and massdiffhist
}
\description{
Bin mass differences from a massdiff object produced by
\code{\link{massdiff}} into a histogram, with counts weighted by the
correlation estimate. The massdiff object must contain the correlation test
results, product of \code{\link{corrPairsMSI}} or
\code{\link{corrPairsMSIchunks}} functions.
}
\seealso{
\code{\link{massdiff}} to generate the mass difference list

\code{\link{plot.massdiffhist}} for plotting and annotating the
         resulting object
}
