% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.msimat.R
\name{identify.msimat}
\alias{identify.msimat}
\title{Identify peak on mass spectrum plot}
\usage{
\method{identify}{msimat}(d, plot = TRUE, give.names = TRUE, ...)
}
\arguments{
\item{d}{msimat object used to plot the mass spectrum}

\item{plot}{logical; display label beside identified peak with the peak mass
value?}

\item{give.names}{logical; return peak masses instead of their indices}

\item{...}{Other arguments to \code{\link{identify}}}
}
\value{
Vector of peak masses if \code{give.names=TRUE}, otherwise indices
        for the identified peaks
}
\description{
Interactively identify peak on mass spectrum plot produced by
\code{\link{plot.msimat}}. Click on the plot to identify the closest mass
peak. Right-click to stop on most systems.
}
\seealso{
\code{\link{plot.msimat}}
}
