% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.msimat.R
\name{image.msimat}
\alias{image.msimat}
\title{Plot mass spectrometry imaging image of a msimat object}
\usage{
\method{image}{msimat}(d, peak = NULL, nrow = NULL, ncol = NULL,
  col = grDevices::terrain.colors(36), main = NULL, ...)
}
\arguments{
\item{d}{msimat; MSI dataset imported with \code{\link{msimat}}}

\item{peak}{Name of the peak to plot. If NULL, then plot the total intensity
for all peaks}

\item{nrow}{Number of rows in the MSI image matrix}

\item{ncol}{Number of columns in the MSI image matrix}

\item{col}{Color scale to use (passed to \code{\link{image}})}

\item{main}{Title of plot, otherwise the peak number}

\item{...}{Other parameters to pass to \code{\link{image}}}
}
\description{
From an MSI dataset imported with \code{\link{msimat}}, plot an image of the
signal intensity for a specified peaks. The image dimensions must be
specified as either \code{nrow} or |code{ncol}.
}
