% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass2adduct.R
\docType{package}
\name{mass2adduct-package}
\alias{mass2adduct}
\alias{mass2adduct-package}
\title{MassSpec Mass Difference Finder}
\description{
Tools for calculating mass differences in mass spectrometry data to explore
molecular adducts. To accompany Janda et al. (MS in prep.)
}
\details{
Import mass spectrometry imaging intensity data with \code{\link{readMSI}}.

Tabulate mass differences from simple list of mass peaks with
\code{\link{diffTabulate}}, or from imported MSI dataset with
\code{\link{diffTabulateMSI}}.

Bin mass differences into histogram with \code{\link{diffHist}}.

Get pairs of mass peaks corresponding to a specific mass difference with
\code{\link{diffGetPeaks}}.

Find the closest-matching mass differences for a set of known molecular
adducts with \code{\link{adductMatch}}; conversely, find the closest-
matching known molecular adducts for the most-abundant mass differences
with \code{\link{topAdducts}}.

Find significant correlations between specified pairs of mass peaks in
mass spectrometry imaging intensity data using \code{\link{corrPairsMSI}};
or find correlations between a single ion of interest vs. all other peaks
using \code{\link{corrSinglePeakMSI}}.

Accompanying data: List of common molecular \code{\link{adducts}} and their masses
}
\author{
\strong{Maintainer}: Brandon Seah \email{kbseah@mpi-bremen.de}

Authors:
\itemize{
  \item Moritz Janda
  \item Manuel Liebeke \email{mliebeke@mpi-bremen.de}
}

}
