% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msimatreadcsv.R
\name{msimatreadcsv}
\alias{msimatreadcsv}
\title{Read CSV file to msimat object}
\usage{
msimatreadcsv(csv, sep, remove.zeroes, header, check.names, row.names, ...)
}
\arguments{
\item{csv}{filename; CSV-formatted table of intensities arranged by peak masses
(columns, with column headers) and pixel spots (rows, with row names)}

\item{sep}{character; column separator in CSV file (default: ",")}

\item{remove.zeroes}{logical; remove columns where all values are zero? (default: TRUE)}

\item{header}{logical; (default: TRUE)}

\item{check.names}{logical; (default: FALSE)}

\item{row.names}{numeric; column in import table specifying row names (default: 1)}

\item{...}{Parameters to be passed to read.csv()}
}
\value{
Object of class msimat
}
\description{
Internal function used by \code{\link{msimat}} to read CSV file
}
\keyword{internal}
