% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msimatreadtriplet.R
\name{msimatreadtriplet}
\alias{msimatreadtriplet}
\title{Read triplet files to msimat object}
\usage{
msimatreadtriplet(rows, cols, vals, peaks, spots)
}
\arguments{
\item{rows}{filename; File of row numbers, 0-based (see note on triplet
representation above), one per line}

\item{cols}{filename; File of column numbers, 0-based, one per line}

\item{vals}{filename; File of intensity values, one per line}

\item{peaks}{filename; File of peak mass values, one per line. Order in list
should correspond to the column numbers given to argument cols}

\item{spots}{filename; File of pixel spot names, one per line. Order in list
should correspond to the row numbers given to argument rows}
}
\value{
Object of class msimat
}
\description{
Internal function used by \code{\link{msimat}} to read triplet files
}
\keyword{internal}
