% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.massdiffhist.R
\name{plot.massdiffhist}
\alias{plot.massdiffhist}
\title{Plot massdiffhist histogram with named adducts labeled}
\usage{
\method{plot}{massdiffhist}(hist, add = adducts, labels = NULL, pch = 0,
  col = "blue", pos = 4, main = "Mass difference histogram",
  xlab = "mass difference (m/z)", ...)
}
\arguments{
\item{add}{data.frame; adduct data to use to annotate the mass diff peaks,
(default: built-in data set \code{adducts})}

\item{labels}{numeric; label histogram peaks with top N adducts from the
adduct table specified to option \code{add} (default: NULL - no labels)}

\item{pch}{Plot character to use for highlighting adduct peaks}

\item{col}{Color for labels and text}

\item{pos}{Position of text label, values 1, 2, 3, 4 are below, left, above,
and to the right of points respectively.}

\item{main}{character; Title for histogram plot}

\item{xlab}{character; X-axis label}

\item{...}{Other parameters to pass to plot}

\item{hist;}{Histogram of a massdiff object produced by
\code{\link{hist.massdiff}}}
}
\description{
Plot massdiff histogram with option of labeling peaks corresponding to known
adducts or chemical transformations. The best-matching adducts for each
histogram bin are found internally with \code{\link{adductMatch.histogram}}.
}
