% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.msimat.R
\name{points.msimat}
\alias{points.msimat}
\title{Points method for object msimat}
\usage{
\method{points}{msimat}(d, type = "h", col = "blue", ...)
}
\arguments{
\item{d}{msimat; MSI imaging object created by function \code{\link{msimat}}}

\item{type}{What type of plot to be drawn, passed to \code{\link{points}}}

\item{col}{color for plot}

\item{...}{Other parameters to pass to \code{plot}}
}
\description{
Points overlay of graph of total intensity per mass peak
}
\seealso{
\code{\link{plot.msimat}}
}
