% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsAdducts.R
\name{pointsAdducts}
\alias{pointsAdducts}
\title{Overlay adduct parental and adduct masses on plot of msimat object}
\usage{
pointsAdducts(d, diff, which = "adduct", signif = TRUE, pch = 20,
  cex = 0.5, col = "red", ...)
}
\arguments{
\item{d}{msimat; Object of class msimat created by function \code{\link{readMSI}}}

\item{diff}{data.frame; Mass difference correlation table produced by
function \code{\link{corrPairsMSI}}}

\item{which}{string; Either plot "parental" or "adduct" peaks as overlay}

\item{signif}{logical; Only plot points which have statistically significant
correlation}
}
\description{
Overlay adduct parental and adduct masses on plot of msimat object
}
