% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.massdiff.R
\name{print.massdiff}
\alias{print.massdiff}
\title{Print method for class massdiff}
\usage{
\method{print}{massdiff}(d)
}
\arguments{
\item{d}{massdiff; Object created by function \code{\link{massdiff}}}
}
\description{
Print method for class massdiff
}
