% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.msimat.R
\name{print.msimat}
\alias{print.msimat}
\title{Print method for class msimat}
\usage{
\method{print}{msimat}(d)
}
\arguments{
\item{d}{msimat; Object created by function \code{\link{readMSI}}}
}
\description{
Print method for class msimat
}
