% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMSI.R
\name{readMSI}
\alias{readMSI}
\title{Read mass spectrometry imaging data}
\usage{
readMSI(file, sep = ";", remove.zeroes = TRUE, header = TRUE,
  check.names = FALSE, row.names = 1, ...)
}
\arguments{
\item{file}{filename; CSV export from MSiReader software using Intensity export}

\item{sep}{character; column separator in CSV file (default: ";")}

\item{remove.zeroes}{logical; remove columns where all values are zero? (default: TRUE)}

\item{header}{logical; (default: TRUE)}

\item{check.names}{logical; (default: FALSE)}

\item{row.names}{numeric; column in import table specifying row names (default: 1)}

\item{...}{Parameters to be passed to read.csv()}
}
\value{
Object of class data.frame
}
\description{
Import mass spectrometry imaging data in CSV format, exported from MSiReader
using the Intensity export function. Note that the column headers should be
ion masses, and that table values should be all numeric or there will be an
error.
}
