% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractPeaks.msimat.R
\name{subtractPeaks.msimat}
\alias{subtractPeaks.msimat}
\title{Remove peaks from an msimat object within specified mass accuracy}
\usage{
\method{subtractPeaks}{msimat}(input, vec, ppm1, ppm2)
}
\arguments{
\item{input}{numeric; msimat object containing peaks to be removed}

\item{vec}{numeric; List of peaks to be removed from d}

\item{ppm1}{numeric; Mass accuracy for peaklist v1, in ppm}

\item{ppm2}{numeric; Mass accuracy for peaklist v2, in ppm}
}
\value{
msimat object
}
\description{
Remove peaks, e.g. known contaminants, from an msimat object. The peaks are
specified by their m/z values, and peak intensities are not taken into
account. Mass accuracies (expressed as ppm values) for both the msimat
dataset and the removal peaklist must be specified, and will be used to
determine the margin of error for excluding a given peak.
}
\seealso{
\code{\link{subtractPeaks.numeric}}

\code{\link{filterPeaks}}
}
