% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractPeaks.numeric.R
\name{subtractPeaks.numeric}
\alias{subtractPeaks.numeric}
\title{Remove peaks from a peaklist within specified mass accuracy}
\usage{
\method{subtractPeaks}{numeric}(input, vec, ppm1, ppm2, index = TRUE)
}
\arguments{
\item{input}{numeric; Peaklist to be subtracted from}

\item{vec}{numeric; List of peaks to be removed from v1}

\item{ppm1}{numeric; Mass accuracy for peaklist v1, in ppm}

\item{ppm2}{numeric; Mass accuracy for peaklist v2, in ppm}

\item{index}{logical; Report indices of peaks to be retained if TRUE, else
report values}
}
\value{
indices of vector v1 to be retained if \code{index=TRUE}, otherwise
        the values of vector v1 corresponding to those indices.
}
\description{
Remove peaks from a peaklist within specified mass accuracy
}
