% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.massdiff.R
\name{summary.massdiff}
\alias{summary.massdiff}
\title{Summary method for class massdiff}
\usage{
\method{summary}{massdiff}(d)
}
\arguments{
\item{d}{massdiff; Object created by function \code{\link{massdiff}}}
}
\description{
Summary method for class massdiff
}
