% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPeakIntensities.msidf.R
\name{sumPeakIntensities.msidf}
\alias{sumPeakIntensities.msidf}
\title{Sum total pixel intensities per peak and report as data frame}
\usage{
\method{sumPeakIntensities}{msidf}(d)
}
\arguments{
\item{d}{msidf; MSI data imported with \code{\link{readMSI}}}
}
\value{
data.frame of peaks and intensities
}
\description{
Sum total pixel intensities per peak and report as data frame
}
