% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPeakIntensities.msimat.R
\name{sumPeakIntensities.msimat}
\alias{sumPeakIntensities.msimat}
\title{Sum total pixel intensities per peak and report as data frame}
\usage{
\method{sumPeakIntensities}{msimat}(d)
}
\arguments{
\item{d}{msimat; MSI data imported with \code{\link{readMSI}}}
}
\value{
data.frame of peaks and intensities
}
\description{
Sum total pixel intensities per peak and report as data frame
}
