% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPeakIntensities.msitsm.R
\name{sumPeakIntensities.msitsm}
\alias{sumPeakIntensities.msitsm}
\title{Sum total pixel intensities per peak and report as data frame}
\usage{
\method{sumPeakIntensities}{msitsm}(d)
}
\arguments{
\item{d}{msitsm; MSI data imported with \code{\link{readMSI}}}
}
\value{
data.frame of peaks and intensities
}
\description{
Sum total pixel intensities per peak and report as data frame
}
