% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPeakIntensities.R
\name{sumPeakIntensities}
\alias{sumPeakIntensities}
\title{Sum total pixel intensities per peak and report as data frame}
\usage{
sumPeakIntensities(d)
}
\arguments{
\item{d}{msitsm or msidf; MSI data imported with \code{\link{readMSI}}}
}
\value{
data.frame of peaks and intensities
}
\description{
Sum total pixel intensities per peak and report as data frame
}
