% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topAdducts.R
\name{topAdducts}
\alias{topAdducts}
\title{Find top adduct candidates from a mass difference histogram}
\usage{
topAdducts(hist, add = adducts, n = 20, use.bw = TRUE, threshold = NULL)
}
\arguments{
\item{hist}{histogram of a massdiff object}

\item{add}{data.frame of adducts (default: "adducts" dataset of mass2adduct)}

\item{n}{numeric; number of top hits to report (default 20)}

\item{use.bw}{logical; adduct is reported as match only if it falls in the
same histogram bin (default: TRUE)}

\item{threshold}{numeric; maximum distance between adduct mass and mass
difference peak to report adduct as a match. use.bw must be FALSE
(default: NULL)}
}
\value{
data.frame of mass bins, counts, closest-matching adduct if known
}
\description{
Identify mass differences with the most observations and report closest-
matching known adducts. This also reports mass differences which do not have
a close match to a known adduct, unlike \code{\link{adductMatch}}
}
\seealso{
\code{\link{adductMatch}} to find the closest-matching mass
         differences for a list of known adducts (the converse of the
         current function)
}
