% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatch.massdiff.R
\name{adductMatch.massdiff}
\alias{adductMatch.massdiff}
\title{Find closest-matching known adducts for a set of mass differences}
\usage{
\method{adductMatch}{massdiff}(x, add = adducts, width = 0.001)
}
\arguments{
\item{x}{massdiff; Mass differences produced by \code{\link{diffTabulate}}}

\item{add}{data.frame of adduct masses (default: "adducts" dataset in package)}

\item{width}{numeric; the precision to which to match the mass value (default: 0.01)}
}
\value{
Object of class massdiff, with additional element "matches" reporting
        the closest matches
}
\description{
For use with internal dataset "adducts" and other similar data
}
\seealso{
\code{\link{topAdducts}} to find the closest-matching adducts for
         the mass differences with the highest counts (the converse of the
         current function)
}
