% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatch.massdiff.R
\name{adductMatch.massdiff}
\alias{adductMatch.massdiff}
\title{Report closest-matching adduct type for a list of mass differences}
\usage{
\method{adductMatch}{massdiff}(x, add = adducts, ppm = 2)
}
\arguments{
\item{x}{massdiff; Mass differences produced by \code{\link{massdiff}}}

\item{add}{data.frame of adduct masses (default: "adducts" dataset in package)}

\item{ppm}{numeric; the mass resolution of the original measurement
(default: 2)}
}
\value{
Object of class massdiff, with additional element "matches" reporting
        the closest matches
}
\description{
Given a set of mass differences produced by function \code{\link{massdiff}},
report for each pair of masses the closest-matching molecular transformation
from a list of such potential adduct types. The built-in data sets
\code{\link{adducts}} and \code{\link{adducts2}} are examples of such lists
of potential adducts.
}
\examples{
d <- msimat(csv=system.file("extdata","msi.csv",package="mass2adduct"),sep=";")
d.diff <- massdiff(d) # Calculate mass differences from imported MSI data
d.diff.annotate <- adductMatch(d.diff,add=adducts2) # Report closest matches

}
\seealso{
\code{\link{adductMatch.massdiffhist}} gives an overview for large data
         sets by matching closest adducts to mass differences already binned
         into a histogram.

\code{\link{topAdducts}} to find the closest-matching adducts for
         the mass differences with the highest counts (the converse of the
         current function)
}
