% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductMatch.R
\name{adductMatch}
\alias{adductMatch}
\title{Find closest-matching known adducts for a set of mass differences or histogram thereof}
\usage{
adductMatch(x, add = mass2adduct::adducts, ...)
}
\arguments{
\item{x}{Object of class massdiff from function \code{\link{massdiff}} or
histogram from running \code{\link{hist}} on a massdiff object}

\item{add}{data.frame; Table of known chemical transfomrations and their
respective masses. (default: built-in data set "adducts")}

\item{...}{Other options to pass to underlying methods}
}
\value{
data.frame of mass differences given to argument diff, with an
        additional column reporting the closest matches
}
\description{
For use with internal dataset "adducts" and other similar data
}
\seealso{
\code{\link{adductMatch.massdiffhist}} method for histogram objects

\code{\link{adductMatch.massdiff}} method for massdiff objects

\code{\link{topAdducts}} to find the closest-matching adducts
         for the mass differences with the highest counts (the converse of the
         current function)
}
