% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffGetPeaks.R
\name{diffGetPeaks}
\alias{diffGetPeaks}
\title{Find pairs of mass peaks corresponding to a specific mass}
\usage{
diffGetPeaks(diff, mass = NULL, ppm = 2, width = NULL)
}
\arguments{
\item{diff}{massdiff; output from \code{\link{massdiff}}}

\item{mass}{numeric; mass of putative adduct or parent ion in m/z units}

\item{ppm}{numeric; mass resolution of the original measurements, to
determine the cutoff for the match (default: 2)}

\item{width}{numeric; Instead of a sliding ppm cutoff, use an absolute m/z 
difference cutoff to call matches (default: NULL)}
}
\value{
Subset of the original massdiff object where the mass difference or
        parent ion is close to the specified mass of interest (within the
        specified width)
}
\description{
From a set of all possible mass pairs, find all pairs corresponding to a
specific mass difference, which might represent a molecular transformation
of interest.
}
