% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffGetPeaksIndex.R
\name{diffGetPeaksIndex}
\alias{diffGetPeaksIndex}
\title{Find pairs of mass peaks corresponding to a specific mass difference (putative adduct)}
\usage{
diffGetPeaksIndex(diff, by = "diff", mass = NULL, width = 0.001)
}
\arguments{
\item{by}{string; Subset the mass by mass difference (putative adduct
transformation) ("diff") or by parent ion ("parent")?}

\item{mass}{numeric; mass of putative adduct or parent ion in m/z units}

\item{width}{numeric; range of values to use}

\item{d}{data.frame; output from diffTabulate}
}
\value{
subset of the original data frame d where the mass difference or
        parent ion is close to the specified mass of interest (within the
        specified width)
}
\description{
Find pairs of mass peaks corresponding to a specific mass difference (putative adduct)
}
