% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTabulate.R
\name{diffTabulate}
\alias{diffTabulate}
\title{Tabulate pairwise combinations of mass peaks}
\usage{
diffTabulate(d, override.limit = FALSE)
}
\arguments{
\item{d}{numeric vector representing m/z values of mass peaks}

\item{override.limit}{logical; continue with computation even if vector
length is >5000 (default: FALSE)}
}
\value{
data.frame of all pairs of masses and their respective differences
}
\description{
Takes all pairwise combinations of masses from a simple numeric vector
that represent the mass values from a mass spectrometry data set.
}
\details{
The number of possible pairs grows combinatorially and so the calculation
can take a very long time for more than 5000 mass values. Use the
\code{override.limit} parameter to override the default limit of 5000.
}
\seealso{
\code{\link{diffTabulateMSI}} to tabulate mass differences from
         MS imaging data object imported with \code{\link{readMSI}} function,
         \code{\link{diffHist}} to plot histogram of mass differences
}
