% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTabulate.R
\name{diffTabulate}
\alias{diffTabulate}
\title{Tabulate pairwise combinations of mass peaks}
\usage{
diffTabulate(d)
}
\arguments{
\item{d}{numeric vector representing m/z values of mass peaks}
}
\value{
data.frame of all pairs of masses and their respective differences
}
\description{
Takes all pairwise combinations of masses from a simple numeric vector
that represent the mass values from a mass spectrometry data set.
}
\details{
The number of possible pairs grows combinatorially and the calculation will
fail for >65536 values (the function uses a 16-bit integer internally).
}
\seealso{
\code{\link{diffTabulateMSI}} to tabulate mass differences from
         MS imaging data object imported with \code{\link{readMSI}} function,
         \code{\link{diffHist}} to plot histogram of mass differences
}
