% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.massdiff.R
\name{hist.massdiff}
\alias{hist.massdiff}
\title{Histogram method for massdiff object}
\usage{
\method{hist}{massdiff}(diff, widthFunc = "equal", width = 0.01,
  plot = FALSE, ...)
}
\arguments{
\item{diff}{massdiff; Output from function \code{\link{diffTabulate}},
containing three numeric vectors (A, B, diff) representing mass
peaks and their differences}

\item{widthFunc}{character; function to use to bin mass differences into
histogram. (default "equal", other options to be added)}

\item{width}{numeric; bin width in m/z units (default 0.01)}

\item{plot}{logical; whether histogram graphic should be plotted (default FALSE)}

\item{...}{Other options to be passed to hist()}
}
\value{
Object of class hist
}
\description{
Bin mass differences from a massdiff object produced by
\code{\link{diffTabulate}} into a histogram, and plot if requested.
}
\seealso{
\code{\link{diffTabulate}} to generate the mass difference list
}
