% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msimat.R
\name{plot.msimat}
\alias{plot.msimat}
\title{Plot method for object msimat}
\usage{
\method{plot}{msimat}(d, type = "l", xlab = "m/z",
  ylab = "Total intensity", ...)
}
\arguments{
\item{d}{msimat; MSI imaging object created by function \code{\link{msimat}}}
}
\description{
Plots a graph of total intensity per mass peak
}
