% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsAdducts.R
\name{pointsAdducts}
\alias{pointsAdducts}
\title{Overlay adduct parental and adduct masses on plot of msimat object}
\usage{
pointsAdducts(d, diff, which = c("adduct", "parent"), invert = FALSE,
  signif = TRUE, pch = 20, cex = 0.5, col = "red", ...)
}
\arguments{
\item{d}{msimat; Object of class msimat created by function
\code{\link{msimat}}}

\item{diff}{data.frame; Mass difference correlation table produced by
function \code{\link{corrPairsMSI}}}

\item{which}{string; Either plot "parent" or "adduct" peaks as overlay}

\item{invert}{logical; Overlay the points which are NOT adducts}

\item{signif}{logical; Only plot points which have statistically significant
correlation}

\item{pch}{Plot character to use, passed to \code{points}}

\item{cex}{Character expansion parameter, passed to \code{points}}

\item{col}{Color for overlay points, passed to \code{points}}

\item{...}{Other parameters to pass to \code{points}}
}
\description{
On a mass spectrum plot (from an msimat object), plot points representing
either parental or adduct masses as an overlay. By default only those flagged
as having statistically significant spatial correlation will be plotted. The
massdiff object should have been processed with \code{\link{adductMatch}} and
\code{\link{corrPairsMSI}} to annotate with adduct matches and correlation
test results.
}
\seealso{
\code{\link{corrPairsMSI}} calculate correlations for mass pairs

\code{\link{massdiff}} tabulate all possible mass pairs
}
