% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsAdducts.R
\name{pointsAdducts}
\alias{pointsAdducts}
\title{Overlay adduct parental and adduct masses on plot of msimat object}
\usage{
pointsAdducts(d, diff, which = c("adduct", "parental"), signif = TRUE,
  pch = 20, cex = 0.5, col = "red", ...)
}
\arguments{
\item{d}{msimat; Object of class msimat created by function
\code{\link{msimat}}}

\item{diff}{data.frame; Mass difference correlation table produced by
function \code{\link{corrPairsMSI}}}

\item{which}{string; Either plot "parental" or "adduct" peaks as overlay}

\item{signif}{logical; Only plot points which have statistically significant
correlation}
}
\description{
On a mass spectrum plot (from an msimat object), plot points representing
either parental or adduct masses as an overlay.
}
\seealso{
\code{\link{corrPairsMSI}} calculate correlations for mass pairs

\code{\link{massdiff}} tabulate all possible mass pairs
}
